+++ Sources/FinnhubSwift/Models/CompanyProfile.swift
@@ -5,10 +5,10 @@ public struct CompanyProfile: Mappable {
     public var currency: String
     public var exchange: String
     public var ipo: String
-    public var marketCapitalization: Double
+    public var marketCapitalization: Float
     public var name: String
     public var phone: String
-    public var shareOutstanding: Double
+    public var shareOutstanding: Float
     public var ticker: String
     public var weburl: String
     public var logo: String
