+++ Sources/FinnhubSwift/Models/NewsSentiment.swift
@@ -2,20 +2,20 @@ import Foundation
 
 public struct Buzz: Mappable {
     public var articlesInLastWeek: Int
-    public var buzz: Double
-    public var weeklyAverage: Double
+    public var buzz: Float
+    public var weeklyAverage: Float
 }
 
 public struct Sentiment: Mappable {
-    public var bearishPercent: Double
-    public var bullishPercent: Double
+    public var bearishPercent: Float
+    public var bullishPercent: Float
 }
 
 public struct NewsSentiment: Mappable {
     public var buzz: Buzz
-    public var companyNewsScore: Double
-    public var sectorAverageBullishPercent: Double
-    public var sectorAverageNewsScore: Double
+    public var companyNewsScore: Float
+    public var sectorAverageBullishPercent: Float
+    public var sectorAverageNewsScore: Float
     public var sentiment: Sentiment
     public var symbol: String
 }
