+++ Sources/FinnhubSwift/Models/Trade.swift
@@ -1,8 +1,8 @@
 import Foundation
 
 public struct Trade: Mappable {
-    public var p: Double // Last price
+    public var p: Float // Last price
     public var s: String // Symbol
     public var t: Int // UNIX milliseconds timestamp
-    public var v: Double // Volume
+    public var v: Float // Volume
 }
